Add files via upload
Added UCI configuratio option to address The NNUE "Halo Effect". The problem is Networks might look at a 
wild position and see a massive, non-material advantage (like a powerful attack) and completely ignore the fact that 
it has to give up a lot of pieces to get it. It believes its attack is so strong that the sacrifice is worth it, even 
if, deep down, it might not be, spotting subtle tactical and positional advantages, but can get a little too aggressive 
or "speculative." So I want to mix the material score with the NNUE score to create a good hybrid solution.

This modification implements a hybrid evaluation that mixes the NNUE score with the material score to mitigate the NNUE 
"Halo Effect". The material score is blended in at a ratio of 1/8, meaning the final score is (7/8 * Score_NNUE) + (1/8 * Score_Material)